feat(Go): add correct bucket name to be used
